requirements.txt
Flask
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
# Optional: gunicorn # For production deployments
fastapi
uvicorn[standard]
pydantic
httpx